\name{altmetrics}
\alias{altmetrics}
\title{Plot altmetrics}
\usage{
  altmetrics(diggname = list("cboettig", "lo0ol"),
    gitid = list("cboettig", "holman"), gittype = "user")
}
\arguments{
  \item{diggname}{Digg username}

  \item{gitid}{GitHub username}

  \item{gittype}{GitHub type, 'user' or 'org'}
}
\description{
  This function plots metrics on persons or articles from
  many sources.
}
\author{
  Scott Chamberlain
}

