\name{gitstats}
\alias{gitstats}
\title{Plot github stats}
\usage{
  gitstats(id = "hadley", type = "user")
}
\arguments{
  \item{id}{name of the github user or organization}

  \item{type}{either "user" or "org"}
}
\description{
  This function plots the total number of forks and
  followers for all repositories of a certain user or
  organization.
}
\author{
  Scott Chamberlain
}

