\name{knithtml}
\alias{knithtml}
\title{Knit to html converter}
\usage{
  knithtml(base64text, ...)
}
\arguments{
  \item{base64text}{knitr string}

  \item{...}{arguments passed on to markdownToHTML}
}
\value{
  HTML string
}
\description{
  Use knitr and markdown to convert knitr to html
}
\author{
  Jeroen and Yihui
}

